rename `main` components into `admin-main` and `superadmin-main` for better clarity
