+++ frontend/src/app/core/components/admin/admin-main/admin-main.component.scss
