+++ frontend/src/app/core/components/superadmin/superadmin-main/superadmin-main.component.spec.ts
@@ -1,7 +1,7 @@
 import { waitForAsync, ComponentFixture, TestBed } from '@angular/core/testing';
 import { NoopAnimationsModule } from '@angular/platform-browser/animations';
 
-import { MainComponent } from './main.component';
+import { MainComponent } from './superadmin-main.component';
 
 describe('MainComponent', () => {
   let component: MainComponent;
@@ -9,7 +9,7 @@ describe('MainComponent', () => {
 
   beforeEach(waitForAsync(() => {
     TestBed.configureTestingModule({
-      imports: [NoopAnimationsModule]
+      imports: [NoopAnimationsModule],
     }).compileComponents();
   }));
